Fixes for products API 1. New function to get children of the chosen category 2. Implementation of new logic for products API
Minor fixes to user API - corrected constructor require
statement
