+++ db-interactions/api-products/api-products-interactions.js
@@ -1,4 +1,3 @@
-const getAllChildrenCategories = require("../../utils/categoryHierarchy");
 const connection = require("../db-init");
 
 async function getAllProducts() {
@@ -13,34 +12,6 @@ async function getAllProducts() {
 
         resolve(results);
       });
-    });
-
-    return results;
-  } catch (error) {
-    throw error;
-  }
-}
-async function getProductsByCategory(categoryId) {
-  try {
-    // const allCategories = await getCategoriesList();
-    // const allRequiredCategories = [
-    //   categoryId,
-    //   ...getAllChildrenCategories(allCategories, categoryId),
-    // ];
-
-    const results = await new Promise((resolve, reject) => {
-      connection.query(
-        "SELECT * FROM items WHERE category_id IN (?)",
-        [categoryId],
-        function (err, results, fields) {
-          if (err) {
-            console.error("Error fetching products:", err);
-            reject(err);
-          } else {
-            resolve(results);
-          }
-        }
-      );
     });
 
     return results;
@@ -72,4 +43,4 @@ async function getCategoriesList() {
   }
 }
 
-module.exports = { getAllProducts, getCategoriesList, getProductsByCategory };
+module.exports = { getAllProducts, getCategoriesList };
